refactor(app-dialogs): split AppDialogHost dialogs into own files
Each AppDialog variant now lives in its own file alongside @Preview
composables, mirroring the per-component layout used in core/ui/kit
(AppConfirmDialog, ActiveSessionConflictDialog, …):

- RestoreSuccessDialog.kt — 3 previews (Light/Dark with undo button +
  no-previous-version variant)
- RestoreFailureDialog.kt — 2 previews (Light + Dark, SchemaTooNew)
- UndoRestoreConfirmationDialog.kt — 2 previews (Light + Dark)
- UndoRestoreSuccessDialog.kt — 2 previews (Light + Dark)
- DialogDateFormatter.kt — shared internal formatMediumDate(epochMs),
  pulled out so both RestoreSuccess and UndoRestoreConfirmation render
  dates the same way.

AppDialogHost.kt keeps only the host wiring (EntryPoint lookup, store
collect, callback bridging) + the AppDialogHostContent dispatch table
the existing AppDialogHostContentTest already drives. The host's
private dialog composables and the inline formatMediumDate helper are
removed; visibility on the extracted dialogs is internal so the host
and tests can call them.
